Kosmetisk: 1 komponent per fil
