+++ src/GjenbruksElement/EksternLenke.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Launch } from "@material-ui/icons/";
+
+const EksternLenke = ({ href, children }) => (
+  <a href={href}>
+    {children} <Launch style={{ width: 16, verticalAlign: "middle" }} />
+  </a>
+);
+
+export default EksternLenke;
