+++ src/GjenbruksElement/WikiMarkdown.js
@@ -1,29 +1,14 @@
 import React, { useEffect, useState } from "react";
 import Markdown from "markdown-to-jsx";
-import { Link } from "react-router-dom";
-import { Launch } from "@material-ui/icons/";
+import Lenke from "./Lenke";
 
 const baseurl =
   "https://raw.githubusercontent.com/wiki/Artsdatabanken/nin-kart/";
 
-const ExternalLink = ({ href, children }) => (
-  <a href={href}>
-    {children} <Launch style={{ width: 16, verticalAlign: "middle" }} />
-  </a>
-);
-
-const Anchor = ({ href, children }) => {
-  href = href.replace("https://nin.artsdatabanken.no", "");
-  const title = children.join(", ");
-  if (href.indexOf("http") === 0)
-    return <ExternalLink href={href}>{title}</ExternalLink>;
-  return <Link to={href}>{title}</Link>;
-};
-
 const options = {
   overrides: {
     a: {
-      component: Anchor
+      component: Lenke
     }
   }
 };
